feat:  Remove AuthRouter, restructure database and add ListUsersUseCase
Removed the AuthRouter file and made several changes related to database repository and connection files. Added a new feature 'ListUsersUseCase' to list all users. Additionally, files in the Shared/Database folder were moved to Shared/Infrastructure for better organization. A new interface for the MikroORMInstance was also introduced in the restructuring process.